backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-Mail
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.14.0
bcrypt
